Rename TribunalOrderOrRequestDetailsHelper
